# Add igraph_random_edge_walk()