--- a/doc/visitors.xxml
+++ b/doc/visitors.xxml
@@ -19,6 +19,7 @@
 
 <section><title>Random walks</title>
 <!-- doxrox-include igraph_random_walk -->
+<!-- doxrox-include igraph_random_edge_walk -->
 </section>
 
 </chapter>