--- a/include/igraph_paths.h
+++ b/include/igraph_paths.h
@@ -133,6 +133,13 @@ DECLDIR int igraph_random_walk(const igraph_t *graph, igraph_vector_t *walk,
                 igraph_integer_t steps,
                 igraph_random_walk_stuck_t stuck);
 
+DECLDIR int igraph_random_edge_walk(const igraph_t *graph,
+                            const igraph_vector_t *weights,
+                            igraph_vector_t *edgewalk,
+                            igraph_integer_t start, igraph_neimode_t mode,
+                            igraph_integer_t steps,
+                            igraph_random_walk_stuck_t stuck);
+
 __END_DECLS
 
 #endif