--- a/tests/visitors.at
+++ b/tests/visitors.at
@@ -31,3 +31,7 @@ AT_KEYWORDS([igraph_bfs bfs breadth-first visitor])
 AT_COMPILE_CHECK([simple/igraph_bfs2.c], [simple/igraph_bfs2.out])
 AT_CLEANUP
 
+AT_SETUP([Random walk (igraph_random_edge_walk):])
+AT_KEYWORDS([igraph_random_edge_walk random_walk])
+AT_COMPILE_CHECK([simple/igraph_random_walk.c])
+AT_CLEANUP