This pull request adds the function `igraph_random_edge_walk()`.

It is analogous to `igraph_random_walk()` with the following differences:

 - it returns a list of edge indices
 - it supports edge weights

Related bug: #981 

There I asked if `random_walk()` should be changed to return an edge list instead of a vertex list. It is probably best to have both variants. Obtaining an edge list from a vertex list is not possible when multiple edges are present between the same vertices.  Obtaining a vertex list from an edge list is possible, but it is inconvenient for undirected graphs (because it is not clear in which direction the edge was traversed).

Still missing:

 - [x] Unit tests
 - [ ] Update `igraph_random_walk()` to also handle weights

I will do these after the review.