@szhorvat 2017-11-28
Changes in the last two commits:

 - fix a bug (wrong result when `mode != IGRAPH_OUT` or the graph is undirected)
 - add benchmark
 - attempt to improve performance

About the performance optimization:

In the case of a weighted graph, we need to select an edge to traverse next with a probability proportional to its weight. This is done by constructing a cumulative distribution vector for the incident edges.  In the previous iteration, this was done in every single step of the algorithm.  This new version will cache and re-use the result. This takes an extra O(|E|) memory.

However, the performance improvement isn't clear:

 - on a small graph, the timing improved from 1.5 s to 1.0 s.
 - on a big graph, the timing increased from 5.4 to 6.0 s.

I suspect that this is due to caching effect.

A further possible improvement would be using a lazy incidence list instead of an incidence list. In practice, this slows down the program (5.4 s -> 7.2 s with original algorithm), possibly because the lazy incidence list uses `double` vectors instead of integer ones.

Finally, due to many conditionals in the main loops, branch prediction failures may have a performance impact too. I am not sufficiently familiar with these effects to be able to guess at performance without testing, so I am going to stop here.