@szhorvat 2018-01-04
Would a similar implementation be okay for the weighted case of `igraph_random_walk()`?  I was thinking of leaving the current implementation of `igraph_random_walk()` mostly alone (other than adding a weights argument), and just adding a second (non-public) function that will be called for the weighted case.