@ntamas 2018-03-06
Re your last comment above (which I have missed somehow): yes, that's okay, just add another non-public function for the weighted case.