@ntamas 2018-03-06
Shall I merge the current PR or will you do it in the same PR?